Move line by line matching to new lib/text.py module
The functionality will also be used for file content matching.
